requirements.txt
streamlit
pandas
Pillow
plotly
openpyxl
scikit-learn
joblib
matplotlib
xgboost
torch
graphviz